+++ HISTORY.md
@@ -1,5 +1,5 @@
 ## v1.0.1
-* Added Cloudflare Turnstile `TestingKeysAndSecrets`.
+* Added Cloudflare Turnstile `TestingSiteKeys` and `TestingSecretKeys`.
 
 ## v1.0.0
 * Initial release.